Version 00.02.05 for FPath_Ex003
This code is designed to support FPath experiment 004. It features the image recognition of colored circles, an efficient algorithm to find the closest point of a specific color on a bitmap and a fast  bitmap container which uses direct memory access to the pixel data rather than GDI calls. Also implemented is a WMF transform to composite a bitmap image on top of a video frame. C# and Windows Media Foundation are used throughout