Version 00.02.07 for FPath_Ex005
This code is designed to support the 2D area filling using a technique dubbed Stigmergic Fill. A new WMF transform has been implemented and this code uses various fast bitmaps to overlay, in real-time, an image frame with a virtual area which clearly shows which pixels of a certain color underlie it. Since the pixels on the areas which have not been filled are clearly a different color it is very simple algorithmically to determine which areas remain to be processed.